convert scaleSlider to functiona along with slider key update with values
